classify-my-bike/app/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.10.2+cpu
torchvision==0.11.3+cpu
pillow
requests
